Seed currencies with no new conversion feed and dedup the seeding helper
