+++ Engine/Setup/BaseSetupHandler.cs
@@ -90,22 +90,7 @@ namespace QuantConnect.Lean.Engine.Setup
             universeSelection.EnsureCurrencyDataFeeds(SecurityChanges.None, seedNewCurrencies: false);
 
             // now set conversion rates
-            Func<Cash, bool> cashToUpdateFilter = currenciesToUpdateWhiteList == null
-                ? (x) => x.CurrencyConversion != null && x.ConversionRate == 0
-                : (x) => currenciesToUpdateWhiteList.Contains(x.Symbol);
-            var cashToUpdate = algorithm.Portfolio.CashBook.Values.Where(cashToUpdateFilter).ToList();
-
-            var securitiesToUpdate = cashToUpdate
-                .SelectMany(x => x.CurrencyConversion.ConversionRateSecurities)
-                .Distinct()
-                .ToList();
-
-            AlgorithmUtils.SeedSecurities(securitiesToUpdate, algorithm);
-
-            foreach (var cash in cashToUpdate)
-            {
-                cash.Update();
-            }
+            AlgorithmUtils.SeedCurrencyConversionRates(algorithm, currenciesToUpdateWhiteList);
 
             Log.Trace($"BaseSetupHandler.SetupCurrencyConversions():{Environment.NewLine}" +
                 $"Account Type: {algorithm.BrokerageModel.AccountType}{Environment.NewLine}{Environment.NewLine}{algorithm.Portfolio.CashBook}");
